+++ io-service/src/main/java/com/example/ioservice/dto/ReviewDTOs/ReviewDTO.java
@@ -1,0 +1,11 @@
+package com.example.ioservice.dto.ReviewDTOs;
+
+import com.example.ioservice.enums.ReviewRatingsEnum;
+
+public record ReviewDTO(Long reviewId,
+                        Long userId,
+                        Long productId,
+                        String title,
+                        String description,
+                        ReviewRatingsEnum reviewRatingsEnum) {
+}