+++ io-service/src/main/java/com/example/ioservice/enums/Role.java
@@ -1,0 +1,7 @@
+package com.example.ioservice.enums;
+
+public enum Role {
+    USER,
+    SELLER,
+    ADMIN
+}