+++ market-service/src/main/java/com/example/marketservice/enums/ReviewRatingsEnum.java
@@ -1,0 +1,9 @@
+package com.example.marketservice.enums;
+
+public enum ReviewRatingsEnum {
+    TERRIBLE,
+    POOR,
+    OK,
+    GOOD,
+    EXCELLENT
+}